cloudfunctions/PriceListDownload/requirements.txt
google-cloud-storage==1.13.0
google-cloud-logging==1.8.0
requests==2.18.4